feat : 스토리 삭제시 confirm 진행

+++ back/src/services/storyPostService.js
@@ -178,18 +178,21 @@ class storyPostService {
     }
   }
 
-  static async deleteUploadImage({ storyId }) {
+  static async deleteUploadedImage({ storyId }) {
     const story = await storyPostModel.findOneByStoryId({ storyId });
     const uploadImage = await imageModel.findOneByImageId({
       imageId: story.thumbnail,
     });
-    const uploadImagePath = uploadImage.path;
-    if (fs.existsSync(uploadImagePath)) {
-      fs.unlinkSync(uploadImagePath);
+    if (uploadImage) {
+      const uploadImagePath = uploadImage.path;
+      if (fs.existsSync(uploadImagePath)) {
+        fs.unlinkSync(uploadImagePath);
+      } else {
+        console.log('File does not exist, so not deleting.');
+      }
     } else {
-      console.log('File does not exist, so not deleting.');
     }
-  }
+}
 }
 
 export { storyPostService };
